refactor: fix file naming convention
