+++ packages/elasticsearch-client/src/createIndex.ts
