+++ packages/elasticsearch-client/src/deleteIndex.ts
