+++ packages/elasticsearch-client/src/getIndexConfig.ts
