+++ packages/elasticsearch-client/src/index.ts
@@ -18,4 +18,4 @@ export { getClient } from './client';
 
 // These types are enums so they need to be exported here in
 // addition to index.d.ts
-export { IndexTypes, ConfigIds } from './get-index-config';
+export { IndexTypes, ConfigIds } from './getIndexConfig';
