+++ packages/elasticsearch-client/src/indexDocument.ts
