+++ packages/elasticsearch-client/src/indexDocumentBulk.ts
